sass: Remove a harmful override on the FA vertical-alignment property
Using the box middle for aligning FA icons tends to result in some
subtle alignment issues that are only noticeable with some browsers in
some display configurations. The only exception to this is the GitHub
icon on the big button at the end of the home page, which absolutely
needs to be middle-aligned due to its size and design.