Redesign the searchbox and allow it to have an advanced search icon
The new design has a color scheme that is more in line with Wesmere's
design without making it seem like someone punched a hole into the
navigation bar, uses an actual <button> for the search icon (meaning it
can be clicked to submit the search form, woohoo!), and has the option
to link to a full-fledged search page, to be used by the forums.

I've also refined the hover/focus highlight effect so it isn't as
distracting as before.

A fortunate side-effect of this change is that the measurements for the
elements of the search form are better defined by a single element
instead of all over the place. It also ensures some cross-browser
consistency for the forums stylesheet supplement by taking some
particularly relevant bits normalize.css that aren't present in phpBB
3.0 and especially concern WebKit-based browsers like Chrome.

Finally, the label for the search box now says "wiki" instead of site,
since that's really what it is. We don't have the means to search both
the wiki and the forums from the same location at this time, and
probably never will.

Due to the scope of these changes, we're going to need a CSS version
bump next.