sass/home: Fix banner effect discrepancies between Firefox and other browsers
It seems Firefox tries to be too smart when applying background-position
and background-size to a multi-layer background and guesses correctly
that they should be applied to the bottom (last) layer and not others.
Other browsers don't appear to do this and instead expect us to redefine
both properties to values that make more sense for that scenario.