home: Add support for hiding entire download entries or just their version
The latter will be needed to implement the iOS version since there's no
point in displaying its version on the Download button -- plus, it'll be
wrong (1.12.6 instead of 1.13.10) since the button currently feeds from
the recommended branch's file list and there's no way to set version
numbers independently for each file.

This is a really hacky solution, but it'll do. Sorry to anyone who might
be reading this, I'm probably disappointing them right now.