sass/forums: Fix whitespace handling issues with the phpBB 3.2 text parser
The new parser produces better markup at the cost of breaking existing
bbCode that relied on the gratuituous linebreaks and whitespace
generated by phpBB 3.1 and earlier, so we just got to adapt to that and
write better templates and CSS too.

This in particular fixes missing margins around the release announcement
blocks, as well as a missing space preceding the expandable block status
chevron.