mw: Use a shorter placeholder for the search box
As of MW 1.31, the placeholder is a string along the lines of "Search
<wiki name>". The problem is that in our wiki, the wiki name is overlong
("The Battle for Wesnoth wiki") so most of it is ellipsized.

I don't see any point in specifying what the search function is for in
the placeholder text, but I'm sure someone will complain "YOU SHOULD'VE
SET IT TO SEARCH THIS WIKI" later on.