home: Rearrange Downloads and System Requirements sections
This makes the list of downloads use a grid layout and also moves it
to the right of the branch info/system requirements section if the
viewport is large enough, in order to make better use of the available
horizontal space.

It also prettifies the "Foo (bar)" platform labels a bit on the icon
list so that the "(bar)" portion is always on a separate line and in a
smaller font size.

Finally, it rephrases the items in System Requirements slightly for
conciseness and adds a few icons to make the OS listings look a little
less intimidating/textwall-y.